Add anything.el like multiple action to selector
Fixed prefix argument 0 bug (when user set
prefix argument key to "" and press the key
which are not recognized by Firefox like <Windows>,
the prefix argument 0 are added.

Fixed the bug when user set Firefox shortcut key
to the supension switch key, the preventDefault()
is missed and Firefox's default behavior is not prevented.