Fix ReferenceError in prompt.js introduced in e681762
let (a = f(a)) { ... } and { let a = f(a); ... } are not equivalent. The
latter case (block + let statement w/ conflicting variable name) is
invalid and raises "Can't access lexical declaration before
initialization" error. Workarounds for this issue are twofold: (1)
rename variables to avoid name confliction or (2) use immediate function
to construct a scope. We take the approach (2) here.