kkk: Fix broken ctrl-tab behavior
The "switch window" functionality is broken on sites where KKK is
enabled. The dialog pops up and multiple tab keypresses are also
recognized but when you release the ctrl-tab keys, nothing happens.
It appears that KKK is eating the "keyup" event.

This commit prevents KKK from hooking into events when ctrl-tab is pressed.